Moved payment data + discount data to separated file
